feat: use server-side event transmit and split backend routes
