+++ backend/Dockerfile
@@ -12,4 +12,4 @@ COPY . .
 
 EXPOSE 1239
 
-CMD ["gunicorn","-w","4","-b","0.0.0.0:1239","--access-logfile", "-", "--error-logfile", "-", "app:app"]
+CMD ["gunicorn","-w","4","-b","0.0.0.0:1239","--worker-class","gevent","--access-logfile", "-", "--error-logfile", "-", "app:app"]
