+++ backend/requirements.txt
@@ -7,7 +7,9 @@ click==8.1.8
 colorama==0.4.6
 configparser==7.1.0
 Flask==3.1.0
+gevent==26.5.0
 gunicorn==23.0.0
+redis==8.0.0
 idna==3.10
 itsdangerous==2.2.0
 Jinja2==3.1.5
